+++ addons/yat/src/commands/builtin/Watch.cs
@@ -4,7 +4,6 @@ using YAT.Attributes;
 using YAT.Enums;
 using YAT.Helpers;
 using YAT.Interfaces;
-using YAT.Overlay.Components.Terminal;
 
 namespace YAT.Commands
 {