+++ CHANGELOG.md
@@ -4,9 +4,19 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+-   Automatic validation of options.
+-   MissingValue method to LogHelper.
+-   Documentation (man command) displays information about options.
+
 ### Changed
 
 -   Update Godot.NET.Sdk version to 4.2.0-beta.6.
+-   Dictionary passed to Execute method can contain both arguments and options.
+-   Command whereami displays abbreviated form by default, adding -l option restores previous behavior.
+-   Methods that generate documentation take no arguments.
+-   Moved documentation on automatic input validation to a separate file.
 
 ## [1.5.0-beta 2023-11-14]